Upgrade to .NET 10, enhance Azure Functions, and improve deployment (#256)
* Update bicep template to use dotnet 10 in isolated mode

Changed hosting to flex consumption

Other improvements:
- No longer uses access keys for storage for functions. Uses managed identity

* Update deployment azure function action.

No longer relies on workaround to deploy using profile since bug was fixed in azure function. Uses managed identity as it should

* feat: Upgrade to .NET 10.0 and update project dependencies

- Updated target framework from net8.0 to net10.0
- Replaced deprecated Azure Functions SDK references with Microsoft.Azure.Functions.Worker and updated related packages.
- Enhanced nullability checks in IssueGateIssues, IssueGateRule, and IssueGateSearch models.
- Added validation for null configurations in IssueGateRulesEvaluator.
- Fixed bug where deployement id was overflowing in an int

* feat: Add Copilot instructions for GitHub Actions gates implementation

+++ src/GitHubActions.Gates.Framework/Models/WebHooks/DeploymentProtectionRuleWebHook.cs
@@ -25,7 +25,7 @@ namespace GitHubActions.Gates.Framework.Models.WebHooks
     public class Deployment
     {
         public string url { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string task { get; set; }
         public string original_environment { get; set; }
@@ -48,7 +48,7 @@ namespace GitHubActions.Gates.Framework.Models.WebHooks
     public class Creator
     {
         public string login { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string avatar_url { get; set; }
         public string gravatar_id { get; set; }
@@ -75,7 +75,7 @@ namespace GitHubActions.Gates.Framework.Models.WebHooks
     [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage()]
     public class Performed_Via_Github_App
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string slug { get; set; }
         public string node_id { get; set; }
         public Owner owner { get; set; }
@@ -114,7 +114,7 @@ namespace GitHubActions.Gates.Framework.Models.WebHooks
     [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage()]
     public class Repository
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string name { get; set; }
         public string full_name { get; set; }
@@ -199,7 +199,7 @@ namespace GitHubActions.Gates.Framework.Models.WebHooks
     public class Owner1
     {
         public string login { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string avatar_url { get; set; }
         public string gravatar_id { get; set; }
@@ -221,7 +221,7 @@ namespace GitHubActions.Gates.Framework.Models.WebHooks
     public class Organization
     {
         public string login { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string url { get; set; }
         public string repos_url { get; set; }
@@ -238,7 +238,7 @@ namespace GitHubActions.Gates.Framework.Models.WebHooks
     public class Sender
     {
         public string login { get; set; }
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
         public string avatar_url { get; set; }
         public string gravatar_id { get; set; }
@@ -260,7 +260,7 @@ namespace GitHubActions.Gates.Framework.Models.WebHooks
     [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage()]
     public class Installation
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string node_id { get; set; }
     }
 }
